♻️ refactor: main page scss 도입
