+++ src/styles/pages/index.scss
@@ -1,0 +1,131 @@
+@import "@styles/common/variables";
+@import "@styles/common/mixins";
+
+.main {
+  .header {
+    width: 1280px;
+    height: 102px ;
+    background-color: black;
+
+    &-wrpper {
+      display: flex;
+      justify-content:  center;
+    }
+  }
+
+  .banner-wrapper{
+    margin-top: 20px;
+    display: flex;
+    justify-content: center;
+  }
+
+  .category {
+    height: 77.5px;
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    flex-grow: 1;
+
+    &-wrapper {
+      margin-top: 50px;
+      display: flex;
+      justify-content: center;
+    }
+
+    &-list {
+      width: 976.39px;
+      display: flex;
+      flex-wrap: wrap;
+      gap:32.5px 1px;
+    }
+
+    &-text{
+      font-family: Noto Sans;
+      font-size: 14px;
+      min-width: 93px;
+      height: 100%;
+      margin-top: 25px;
+      text-align: center;
+    }
+  }
+
+  .newgoods{
+    margin: 0px;
+    list-style: none;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    position: relative;
+
+    &-title{
+      font-size: 26px;
+      font-weight: bold;
+    }
+
+    &-wrapper{
+      padding: 0px;
+      display: flex;
+      justify-content: center;
+    }
+
+    &-list{
+      padding: 0px;
+      padding-inline-start: 0px;
+      width: 999px;
+      display: flex;
+      justify-content: center;
+      gap: 50px 20px;
+      flex-wrap: wrap;
+    }
+
+    & .info {
+      &-top {
+        max-width: 140px;
+        margin-top: 10px;
+        font-size: 14px;
+        white-space: nowrap;
+        overflow: hidden;
+        text-overflow: ellipsis;
+      }
+
+      &-middle{
+        color: $FONT_GRAY_COLOR;
+        margin-top: 7px;
+        font-size: 12px;
+      }
+      &-bottom{
+        font-weight: bold;
+        margin-top: 18px;
+        font-size: 16px;
+      }
+    }
+  }
+
+  .newgoodstext {
+    &-wrapper{
+      margin-top: 50px;
+      display: flex;
+      justify-content: center;
+      margin-bottom: 50px;
+    }
+
+    &-container{
+      width: 999px;
+    }
+  }
+
+
+  @include mobile{
+    .header {
+      width: 100%;
+      height: 102px ;
+
+      &-wrapper{
+        margin:0px;
+        display: flex;
+        justify-content:  center;
+        width:100%;
+      }
+    }
+  }
+}
