--- src/styles/pages/main.scss
@@ -1,141 +1,0 @@
-@import "@styles/common/variables";
-@import "@styles/common/mixins";
-
-
-.header-wrapper{
-    display: flex;
-    justify-content:  center;
-    .header {
-        width: 1280px;
-         height: 102px ;
-         background-color: black;
-    }
-}
-
-
-
-.banner-wrapper{
-    margin-top: 20px;
-    display: flex;
-    justify-content: center;
-    
-}
-
-.banner{
-    width: 996px;
-    height: 329px;
-    background-color: #888888;
-    align-self: center;
-    color:white;
-}
-
-.category-wrapper{
-    margin-top: 50px;
-    display: flex;
-    justify-content: center;
-
-}
-
-.category-list{
-    width: 976.39px;
-    display: flex;
-    flex-wrap: wrap;
-    gap:32.5px 1px;
-}
-
-.category{
-    height: 77.5px;
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    flex-grow:1;
-
-}
-
-.categorytext{
-    font-family: Noto Sans;
-    font-size: 14px;
-    min-width: 93px;
-    height: 100%;
-    margin-top: 25px;
-    text-align: center;
-}
-
-.newgoodstext-wrapper{
-    margin-top: 50px;
-    display: flex;
-    justify-content: center;
-    margin-bottom: 50px;
-
-}
-
-.newgoodstext-container{
-    width: 999px;
-}
-
-.newgoods-title{
-font-size: 26px;
-font-weight: bold;
-}
-
-.newgoods-wrapper{
-    padding: 0px;
-    display: flex;
-    justify-content: center;
-}
-
-.newgoods-list{
-    padding: 0px;
-    padding-inline-start: 0px;
-    width: 999px;
-    display: flex;
-    justify-content: center;
-    gap: 50px 20px;
-    flex-wrap: wrap;
-}
-.newgoods{
-margin: 0px;
-list-style: none;
-display: flex;
-flex-direction: column;
-justify-content: center;
-position: relative;
-
-    .info-top{
-        max-width: 140px;
-        margin-top: 10px;
-        font-size: 14px;
-        white-space: nowrap;
-        overflow: hidden;
-        text-overflow: ellipsis;
-    }
-    .info-middle{
-        color: #999999;
-        margin-top: 7px;
-        font-size: 12px;
-    }
-    .info-bottom{
-        font-weight: bold;
-        margin-top: 18px;
-        font-size: 16px;
-    }
-
-}
-
-@include desktop{
-    
-}
-
-@include mobile{
-    .header-wrapper{
-        margin:0px;
-        display: flex;
-        justify-content:  center;
-        width:100%;
-        .header {
-            width: 100%;
-             height: 102px ;
-             background-color: black;
-        }
-    }
-}
